Add files via upload
This is what I think is that is needed for the Tokenizer to work.  It all compiles together.
I had to flip around a lot of the C++ functions to return Strings instead of statuses (via  int/bool) (since you can't change something you pass into a java function).  Error checking will have to be done through Exceptions.
I realize I did something super inefficient for dealing with the strings, it can eventually be changed to use the StringBuffer class, but for now "if it ain't broke don't fix it", since it compiles to my knowledge.